+++ tests/data/cases/preview_return_annotation_brackets_string.py
@@ -1,4 +1,4 @@
-# flags: --preview
+# flags: --unstable
 # Long string example
 def frobnicate() -> "ThisIsTrulyUnreasonablyExtremelyLongClassName | list[ThisIsTrulyUnreasonablyExtremelyLongClassName]":
     pass